Changes as requested. simAgent line 42, is a mutex required?
